+++ src/X/Environment.java
@@ -6,7 +6,7 @@ import X.Nodes.Type;
 
 public class Environment {
 
-    public static Type booleanType, i64Type, strType, voidType, errorType, f32Type, f64Type, i8Type, charPointerType, i32Type, voidPointerType, variaticType;
+    public static Type booleanType, i64Type, strType, voidType, errorType, f32Type, f64Type, i8Type, charPointerType, i32Type, voidPointerType, variaticType, u8Type, u32Type, u64Type;
     public static Function 
         sin, cos, pow, malloc, free, calloc, realloc, exit, fmod, fabs, printf, __isoc99_scanf, fgets, getchar, memcmp
     ;